Add missing boundary check.
Whenever we'd reach the end of a tunnel for an odd number, we'd
still be 1 block off of the end of the tunnel (as is the behavior
with "tunneldig()". As such, we'd end up with a cube with off-by-one
sizes for every other column.

Also, the calculation for which height to go down upon reaching the
end of the tunnel for odd numbered rows has been subtracted by one
as to not end up creating accidental "stairs".